// libs/sota-common/src/enums/BlockchainPlatform.ts
export enum BlockchainPlatform {
  Bitcoin = 'btc',
  Ethereum = 'eth',
  Ripple = 'xrp',
  BitcoinCash = 'bch',
  BitcoinSV = 'bsv',
  EOS = 'eos',
  Stellar = 'xlm',
  Litecoin = 'ltc',
  Tron = 'trx',
  Cardano = 'ada',
  Dash = 'dash',
  Monero = 'xmr',
  EthereumClassic = 'etc',
  NEO = 'neo',
  Tomo = 'tomo',
  Nem = 'nem',
  Iota = 'iota',
}

export default BlockchainPlatform;